[dagster-airflow 1/2] Split apart test_factory.py
Summary:
This splits up `test_factory.py` in the Airflow tests by deploy target (python, docker, k8s)

Test Plan: unit

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1767
